+++ app/src/main/java/com/example/itgproductapiapp/ui/navigation/SelectedProduct.kt
@@ -2,7 +2,7 @@ package com.example.itgproductapiapp.ui.navigation
 
 data class SelectedProduct (
     val name: String? = null,
-    val price: Double? = null,
+    val price: Int? = null,
     val imageUrl: List<String>? = null,
     val category: String? = null
 )
